==> 1. data_light.ipynb <==
# pip install fastparquet

# 데이터 경량화 (CSV to Parquet)

- 2019-Oct: 5.67GB -> 2.01GB

import pandas as pd
import numpy as np

# 데이터 로드
path1 = '/Users/hyegwan/Folder/cp2/data/2019-Oct.csv'
path2 = '/Users/hyegwan/Folder/cp2/data/2019-Nov.csv'

data_oct = pd.read_csv(path1)
data_nov = pd.read_csv(path2)

data_oct.to_parquet('/Users/hyegwan/Folder/cp2/data/Oct.parquet')
data_nov.to_parquet('/Users/hyegwan/Folder/cp2/data/Nov.parquet')

df1 = pd.read_parquet('/Users/hyegwan/Folder/cp2/data/Oct.parquet')
df2 = pd.read_parquet('/Users/hyegwan/Folder/cp2/data/Nov.parquet')

df1.head(3)

df2.head(3)

# 결측치 처리

df1.isnull().sum()

df2.isnull().sum()

user_session 결측치 삭제

df1.dropna(subset=['user_session'], inplace=True) 
df2.dropna(subset=['user_session'], inplace=True) 

category_code, brand 결측치를 unknown으로 대체

df1.fillna({'category_code':'unknown', 'brand':'unknown'}, inplace=True)
df2.fillna({'category_code':'unknown', 'brand':'unknown'}, inplace=True)

확인

df1.isnull().sum()

df2.isnull().sum()

# event_time 전처리

- UTC 글자 제거
- datetime으로 변경
- GMT+4로 가정하므로 4시간을 더해주기

# UTC 글자 제거
df1['event_time'] = df1['event_time'].apply(lambda x : x[:-4])

# datetime으로 변경
df1['event_time'] = df1['event_time'].astype('datetime64')

# GMT+4 시간대로 변경
import datetime
df1['event_time'] = df1['event_time']+datetime.timedelta(hours=4)

df1.head(3)

df1.to_parquet('/Users/hyegwan/Folder/cp2/data/2019-Oct.parquet')
