==> 2. EDA.ipynb <==
# !pip install fastparquet

# from google.colab import drive
# drive.mount('/content/drive')

import pandas as pd
import numpy as np
import datetime as dt
import warnings

# 2019-Oct-Ver2

df = pd.read_parquet('/content/drive/MyDrive/MyCP2/2019-Oct.parquet')

df['event_type'] = df['event_type'].astype('category')
df['category_code'] = df['category_code'].astype('category')

df1['category_code'] = df1['category_code'].astype('string')
df1.update(df1['category_code'].fillna('NaN.NaN.NaN'))
df1['category_code'] = df1['category_code'].astype('category')

df1['main_category'] = df1['category_code'].str.split('.', expand = True)[0].astype('category')
df1['sub_category'] = df1['category_code'].str.split('.', expand = True)[1].astype('category')
df1['sub_sub_category'] = df1['category_code'].str.split('.', expand = True)[2].astype('category')

df1 = df1.drop(columns = ['category_code'])

df1['brand'] = df1['brand'].astype('string')
df1.update(df1['brand'].fillna('unknown'))
df1['brand'] = df1['brand'].astype('category')

df1['purchased'] = np.where(df1['event_type'] == 'purchase' , 1, 0)
df1['viewed'] = np.where(df1['event_type'] == 'view' , 1, 0)
df1['put_cart'] = np.where(df1['event_type'] == 'cart' , 1, 0)

df2 = df1.astype({'product_id' : 'category', 'category_id' : 'category', 'user_id' : 'category', 'user_session' : 'category'})

df2.to_parquet('/content/drive/MyDrive/MyCP2/2019-Oct-Ver2.parquet')

# ItemsUsers

path = '/content/drive/MyDrive/MyCP2/2019-Oct-Ver2.parquet'
data = pd.read_parquet(path, engine='fastparquet')

df = data.copy()

df.columns

# users_id_num 컬럼 생성
df_users = pd.DataFrame(df['user_id'].drop_duplicates())
df_users = df_users.assign(
    users_id_num=np.arange(len(df_users))
    ).reset_index(drop=True)

# items_id_num 컬럼 생성
df_items = pd.DataFrame(df['product_id'].drop_duplicates())
df_items = df_items.assign(
    items_id_num=np.arange(len(df_items))
    ).reset_index(drop=True)

# id_num merge
df = pd.merge(df, df_users)
df = pd.merge(df, df_items)

# items_users_id_tuple 컬럼 생성
df['items_users_id_tuple'] = list(
    zip(df.users_id_num, df.items_id_num)
    )

# 'purchased', 'user_session' 컬럼 삭제
df = df.drop(['purchased', 'user_session'], axis=1)

# ItemsUsers 파케이로 저장
df.to_parquet('/content/drive/MyDrive/MyCP2/ItemsUsers.parquet')

# Items

path = '/content/drive/MyDrive/MyCP2/2019-Oct-Ver2.parquet'
data = pd.read_parquet(path, engine='fastparquet')

df = data.copy()

# items_id_num 컬럼 생성
df_items = pd.DataFrame(df['product_id'].drop_duplicates())
df_items = df_items.assign(
    items_id_num=np.arange(len(df_items))
    ).reset_index(drop=True)

df = pd.merge(df, df_items)

df['viewed'] = np.where(df['event_type'] == 'view' , 1, 0)
df['put_cart'] = np.where(df['event_type'] == 'cart' , 1, 0)

df.columns

df1 = df[['product_id', 'user_id', 'main_category', 'sub_category', 'sub_sub_category', 'viewed', 'put_cart', 'purchased', 'items_id_num']]

df1.head()

## product_id별 view_num, cart_num, purchase_num 컬럼 생성

# product_id 중복값 삭제
df2 = pd.DataFrame(df1['product_id'].drop_duplicates())

# product_id별 view, cart, purchas 횟수 계산
df3 = pd.DataFrame(df1.groupby(['product_id'])['viewed'].sum()).reset_index()
df4 = pd.DataFrame(df1.groupby(['product_id'])['put_cart'].sum()).reset_index()
df5 = pd.DataFrame(df1.groupby(['product_id'])['purchased'].sum()).reset_index()

df6 = pd.merge(df2, df3)
df7 = pd.merge(df6, df4)
df8 = pd.merge(df7, df5)

df8.rename(columns = {'viewed':'view_num', 'put_cart':'cart_num', 'purchased':'purchase_num'},inplace=True)

df8.head()

# product_id별 카테고리 컬럼 추가
df10 = df[['product_id', 'main_category', 'sub_category', 'sub_sub_category', 'items_id_num']]
df11 = df10.drop_duplicates()

df9 = pd.merge(df8, df11)

df9.head()

# category 하나로 묶기
cols = ['main_category', 'sub_category', 'sub_sub_category']
df9['category'] = df9[cols].apply(lambda row: ','.join(row.values.astype(str)), axis=1)

df9.drop(['main_category', 'sub_category', 'sub_sub_category'], axis=1, inplace=True)

# 예시
df9[df9['category'].str.contains('electronics')]

# items_features 컬럼 생성: (items_id_num, [category])
features = df9[['category']].apply(
    lambda x: ','.join(x.map(str)), axis=1)
features = features.str.split(',')
features = list(zip(df9['items_id_num'], features))

df9['items_features'] = features

df9

df9['items_features'] = df9['items_features'].astype(str)

df9.to_parquet('/content/drive/MyDrive/MyCP2/Items.parquet')

df9

# Users

path = '/content/drive/MyDrive/MyCP2/ItemsUsers.parquet'
data = pd.read_parquet(path, engine='fastparquet')

df = data.copy()

df.columns

df1 = df[['user_id', 'product_id', 'event_type']]

# 같은 user의 같은 item에 대한 중복 view, cart, purchase 삭제
df1 = df1.drop_duplicates()

len(df1)

df1 = df1.groupby(
    ['user_id', 'product_id'])['event_type'].apply(
        ','.join).reset_index()

df1.rename(columns = {'event_type':'events'},inplace=True)

# 예시
df1[df1['events'].str.contains('cart')]

df1.to_parquet('/content/drive/MyDrive/MyCP2/Users.parquet')

path = '/content/drive/MyDrive/MyCP2/Users.parquet'
data = pd.read_parquet(path, engine='fastparquet')

df = data.copy()

df1 = data.copy()

df.head()

df1['events'].unique()

# users_id_num 컬럼 추가
df_users = pd.DataFrame(df['user_id'].drop_duplicates())
df_users = df_users.assign(
    users_id_num=np.arange(len(df_users))
    ).reset_index(drop=True)

df2 = pd.merge(df1, df_users)

# Users_features 컬럼 생성: (users_id_num, [events])
features1 = df2[['events']].apply(
    lambda x: ','.join(x.map(str)), axis=1)
features1 = features1.str.split(',')
features1 = list(zip(df2['users_id_num'], features1))

df2['users_features'] = features1

df2

# 예시
df2[df2['user_id']==566233096]

df2['users_features'] = df2['users_features'].astype(str)

df2.info()

df2.to_parquet('/content/drive/MyDrive/MyCP2/Users.parquet')

df2

df2[df2['user_id']==566233096]
